requirements.txt
Flask==1.1.2
Flask-Admin==1.5.8
Flask-SQLAlchemy==2.5.1
Flask-Mail-SendGrid==0.3
Flask-SocketIO==5.0.1
Flask-Dance==4.0.0
Flask-Security-Too==4.0.1
marshmallow==2.20.2
psycopg2-binary==2.8.6
requests==2.25.1
gevent==21.1.2
gevent-websocket==0.10.1
base58==2.1.0
python-stdnum==1.16
qrcode==6.1
email-validator==1.1.2
